pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phish-email-detection-agent"
version = "0.1.0"
description = "Structured agent app template"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Argis Team" }]

dependencies = [
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
  "openai-agents",
  "litellm",
]

[project.optional-dependencies]
ui = [
  "gradio>=4.0",
]
api = [
  "fastapi>=0.110",
  "uvicorn>=0.30",
]
analysis = [
  "openai>=1.40",
  "pillow>=10.0",
  "pytesseract>=0.3.10",
  "opencv-python-headless>=4.8",
  "pyzbar>=0.1.9",
  "faster-whisper>=1.0",
]
test = [
  "pytest>=7.4",
]
dev = [
  "ruff>=0.3",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100

[tool.setuptools.packages.find]
where = ["src"]
include = ["phish_email_detection_agent*"]

[tool.setuptools.package-data]
phish_email_detection_agent = ["config/*.yaml"]

[project.scripts]
phish-email-detection-agent = "phish_email_detection_agent.__main__:main"
